Reset passages stores when language(s) change
