+++ frontend/src/routes/languages/+page.svelte
@@ -16,6 +16,19 @@
   import MobileLanguageDisplay from './MobileLanguageDisplay.svelte'
   import DesktopLanguageDisplay from './DesktopLanguageDisplay.svelte'
 
+  function handleLangChange(e: Event, lang: string) {
+    const input = e.currentTarget as HTMLInputElement
+    languagesClickedOrderStore.update((arr) => {
+      if (input.checked) {
+        // append only if not already present
+        return arr.includes(lang) ? arr : [...arr, lang]
+      } else {
+        // remove if unchecked
+        return arr.filter((v) => v !== lang)
+      }
+    })
+  }
+
   $: {
     $resourceTypesStore = $resourceTypesStore.filter((item) =>
       $langCodesStore.includes(getResourceTypeLangCode(item))
@@ -39,6 +52,7 @@
 
 <Languages
   message="Select up to 2 languages"
+  {handleLangChange}
   {WizardBreadcrumb}
   {WizardBasket}
   {MobileLanguageDisplay}
